+++ client/.changeset/config.json
@@ -9,7 +9,6 @@
   "ignore": [
     "@wikijump/cm-espells",
     "@wikijump/cm-lang-ftml",
-    "@wikijump/cm-tarnation",
     "@wikijump/codemirror",
     "@wikijump/components",
     "@wikijump/css",
