--- client/modules/cm-tarnation/README.md
@@ -1,3 +1,0 @@
-# cm-tarnation
-
-Adds an alternative parser to Lezer to CodeMirror 6. `cm-tarnation` is regex based, being most similar to the Monaco Editor's Monarch syntax highlighting schema and TextMate grammars.
