--- client/modules/cm-tarnation/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "@wikijump/cm-tarnation",
-  "license": "agpl-3.0-or-later",
-  "description": "An alternative parser for CodeMirror 6.",
-  "version": "0.0.0",
-  "keywords": [
-    "wikijump"
-  ],
-  "private": true,
-  "scripts": {},
-  "type": "module",
-  "main": "src/index.ts",
-  "eslintConfig": {
-    "extends": "../../.eslintrc.js"
-  },
-  "dependencies": {
-    "@lezer/common": "^0.15.5",
-    "@lezer/lr": "^0.15.2",
-    "@wikijump/codemirror": "workspace:*",
-    "@wikijump/util": "workspace:*"
-  }
-}
