--- client/modules/cm-tarnation/src/enums.ts
@@ -1,13 +1,0 @@
-export enum Nesting {
-  /** Special value which indicates that a nested region should be ended. */
-  POP
-}
-
-export enum Wrapping {
-  /** The {@link Node} in this match contains the entirety of the branch. */
-  FULL,
-  /** The {@link Node} in this match begins the branch. */
-  BEGIN,
-  /** The {@link Node} in this match ends the branch. */
-  END
-}
