--- client/modules/cm-tarnation/src/index.ts
@@ -1,2 +1,0 @@
-export * from "./language"
-export * from "./util"
