--- client/modules/cm-tarnation/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "compilerOptions": {
-    "baseUrl": "./",
-    "rootDir": "./"
-  },
-  "extends": "../../tsconfig.json",
-  "include": ["src/**/*", "tests/**/*", "../../misc/types/**/*"]
-}
