+++ client/modules/codemirror/src/misc.ts
@@ -1,5 +1,7 @@
+import type { LRParser, ParserConfig } from "@lezer/lr"
 import { animationFrame, idleCallback } from "@wikijump/util"
 import type { EditorState, Text } from "../cm"
+import { Extension, LanguageDescription, LanguageSupport, LRLanguage } from "../cm"
 
 const encoder = new TextEncoder()
 
@@ -80,4 +82,35 @@ export function underline(color: string) {
     <path d="m0 3 l2 -2 l1 0 l2 2 l1 0" stroke="${color}" fill="none" stroke-width=".7"/>
   </svg>`
   return `url('data:image/svg+xml;base64,${btoa(svg)}')`
+}
+
+export interface CreateLezerLanguageOpts {
+  name: string
+  parser: LRParser
+  configure?: ParserConfig
+  alias?: string[]
+  ext?: string[]
+  languageData?: Record<string, any>
+  extensions?: Extension[]
+}
+
+export function createLezerLanguage(opts: CreateLezerLanguageOpts) {
+  const langDesc = Object.assign(
+    { name: opts.name },
+    opts.alias ? { alias: opts.alias } : {},
+    opts.ext ? { extensions: opts.ext } : {}
+  )
+  const langData = { ...langDesc, ...(opts.languageData ?? {}) }
+
+  const load = function () {
+    const lang = LRLanguage.define({
+      parser: opts.parser.configure(opts.configure ?? {}),
+      languageData: langData
+    })
+    return new LanguageSupport(lang, opts.extensions)
+  }
+
+  const description = LanguageDescription.of({ ...langDesc, load: async () => load() })
+
+  return { load, description }
 }
